+++ sentry-android-sqlite/src/test/java/io/sentry/android/sqlite/SQLiteSpanManagerTest.kt
@@ -5,6 +5,7 @@ import io.sentry.IHub
 import io.sentry.SentryIntegrationPackageStorage
 import io.sentry.SentryOptions
 import io.sentry.SentryTracer
+import io.sentry.SpanDataConvention
 import io.sentry.SpanStatus
 import io.sentry.TransactionContext
 import io.sentry.util.thread.IMainThreadChecker
@@ -100,8 +101,8 @@ class SQLiteSpanManagerTest {
         sut.performSql("sql") {}
         val span = fixture.sentryTracer.children.first()
 
-        assertFalse(span.getData("blocked_main_thread") as Boolean)
-        assertNull(span.getData("call_stack"))
+        assertFalse(span.getData(SpanDataConvention.BLOCKED_MAIN_THREAD_KEY) as Boolean)
+        assertNull(span.getData(SpanDataConvention.CALL_STACK_KEY))
     }
 
     @Test
@@ -114,7 +115,7 @@ class SQLiteSpanManagerTest {
         sut.performSql("sql") {}
         val span = fixture.sentryTracer.children.first()
 
-        assertTrue(span.getData("blocked_main_thread") as Boolean)
-        assertNotNull(span.getData("call_stack"))
+        assertTrue(span.getData(SpanDataConvention.BLOCKED_MAIN_THREAD_KEY) as Boolean)
+        assertNotNull(span.getData(SpanDataConvention.CALL_STACK_KEY))
     }
 }